requirements.txt
tqdm
pyyaml==5.4.1
scipy
opencv-python
tensorboard
setuptools==59.5.0
pytz
packaging
numpy==1.22.0
scikit-learn
six
matplotlib
scikit-image
open3d